test: refactor test package
